following almost working

+++ src/components/mainpage/nestedcomponents/followbutton.js
@@ -5,11 +5,14 @@ export class FOLLOW_BUTTON extends React.Component {
   constructor(props) {
     super(props);
 
+    this.changeFollowedFalse = this.changeFollowedFalse.bind(this);
+    this.changeFollowedTrue = this.changeFollowedTrue.bind(this);
+    this.followUser = this.followUser.bind(this);
+    this.unfollowUser = this.unfollowUser.bind(this);
+
     this.state = {
       followed: false,
     };
-    this.changeFollowedFalse = this.changeFollowedFalse.bind(this);
-    this.changeFollowedTrue = this.changeFollowedTrue.bind(this);
   }
 
   changeFollowedFalse = () => {
@@ -30,13 +33,11 @@ export class FOLLOW_BUTTON extends React.Component {
         .firestore()
         .collection("users")
         .doc(this.props.uniqueUid)
-        .collection("followedUserUids")
+        .collection("followedTweeds")
         .onSnapshot((snapshot) => {
           snapshot.forEach((doc) => {
             if (doc.data().uid === this.props.uid) {
               this.changeFollowedTrue();
-            } else {
-              this.changeFollowedFalse();
             }
           });
         });
@@ -44,48 +45,63 @@ export class FOLLOW_BUTTON extends React.Component {
     isUserFollowed();
   }
 
+  followUser = () => {
+    this.changeFollowedTrue()
+      firebase
+        .firestore()
+        .collection("users")
+        .doc(this.props.uid)
+        .collection("userTweeds")
+        .orderBy("created", "asc")
+        .onSnapshot((snapshot) => {
+          snapshot.forEach((doc) => {
+            firebase
+              .firestore()
+              .collection("users")
+              .doc(this.props.uniqueUid)
+              .collection("followedTweeds")
+              .doc()
+              .set({
+                tweed: doc.data().tweed,
+                username: doc.data().username,
+                uid: doc.data().uid,
+              });
+          });
+        });
+  };
+
+  unfollowUser = () => {
+    this.changeFollowedFalse()
+      firebase
+        .firestore()
+        .collection("users")
+        .doc(this.props.uniqueUid)
+        .collection("followedTweeds")
+        .onSnapshot((snapshot) => {
+          snapshot.forEach((doc) => {
+            if (doc.data().uid === this.props.uid) {
+              firebase
+                .firestore()
+                .collection("users")
+                .doc(this.props.uniqueUid)
+                .collection("followedTweeds")
+                .doc(doc.id)
+                .delete();
+            }
+          });
+        });
+  };
+
   render() {
-    let followUser = () => {
-      firebase
-        .firestore()
-        .collection("users")
-        .doc(this.props.uniqueUid)
-        .collection("followedUserUids")
-        .doc(this.props.uid)
-        .set({
-          uid: this.props.uid,
-        })
-        .then(this.changeFollowedTrue());
-    };
-
-    let unfollowUser = () => {
-      firebase
-        .firestore()
-        .collection("users")
-        .doc(this.props.uniqueUid)
-        .collection("followedUserUids")
-        .doc(this.props.uid)
-        .delete()
-        .then(
-          firebase
-            .firestore()
-            .collection("users")
-            .doc(this.props.uniqueUid)
-            .collection("followerTweeds")
-            .doc(this.props.uid)
-            .delete()
-        )
-        .then(this.changeFollowedFalse());
-    };
-
     return (
       <>
         {this.state.followed ? (
-          <button onClick={unfollowUser}>Unfollow</button>
+          <button onClick={this.unfollowUser}>Unfollow</button>
         ) : (
-          <button onClick={followUser}>Follow</button>
+          <button onClick={this.followUser}>Follow</button>
         )}
       </>
     );
   }
 }
+
